move test to LauncherTest
